refactor: move FormalJoinTester
